refactor definitions
